services/image-captioning/requirements.txt
torch==2.7.0+cu128
torchvision==0.22.0+cu128
--extra-index-url https://download.pytorch.org/whl/cu128
transformers
pillow>=9.0.0
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart 
opencv-python-headless
accelerate>=0.26.0